Split X509 code out of tortls.c
